Fix temp file delete path validation

+++ cpan/delete_file.cgi
@@ -6,8 +6,8 @@ require './cpan-lib.pl';
 &ReadParse();
 $tmp_base = $gconfig{'tempdir'} || &default_webmin_temp_dir();
 foreach $f (split(/\0/, $in{'file'})) {
-	$f =~ /^\Q$tmp_base\E\// || &error($text{'delete_efile'});
-	unlink($f);
+	&is_under_directory($tmp_base, $f) || &error($text{'delete_efile'});
+	unlink($f) if (!&is_readonly_mode());
 	}
 &redirect("");
