+++ software/delete_file.cgi
@@ -5,7 +5,7 @@
 require './software-lib.pl';
 &ReadParse();
 my $tmp_base = $gconfig{'tempdir'} || &default_webmin_temp_dir();
-$in{'file'} =~ /^\Q$tmp_base\E\// || &error($text{'delete_efile'});
-unlink($in{'file'});
+&is_under_directory($tmp_base, $in{'file'}) || &error($text{'delete_efile'});
+unlink($in{'file'}) if (!&is_readonly_mode());
 &redirect("");
